requirements.txt
google-adk
google-genai
pandas
db-dtypes
litellm
toolbox-langchain
langchain
immutabledict
sqlglot
fastapi
google-auth-oauthlib
google-cloud-compute